Allow passing floating point numbers for time durations

+++ src/constants.py
@@ -65,5 +65,5 @@ class Ratelimits:
     """Config related to rate limiting."""
 
     REQUESTS_PER_PERIOD = _get_config("REQUESTS_PER_PERIOD", cast=int, default=3)
-    TIME_PERIOD = _get_config("TIME_PERIOD", cast=int, default=20)  # in seconds
-    COOLDOWN_PERIOD = _get_config("COOLDOWN_PERIOD", cast=int, default=100)  # in seconds
+    TIME_PERIOD = _get_config("TIME_PERIOD", cast=float, default=20)  # in seconds
+    COOLDOWN_PERIOD = _get_config("COOLDOWN_PERIOD", cast=float, default=100)  # in seconds
